+++ SurvivalGame/Assets/UI/DraggableUIItem.cs
@@ -26,7 +26,6 @@ public class DraggableUIItem : MonoBehaviour, IBeginDragHandler, IDragHandler, I
 
     public void OnEndDrag(PointerEventData eventData)
     {
-        // NB: At this point the parent could have changed due to InventoryUISlot.OnDrop()
         this.transform.SetParent(parent);
         this.GetComponent<Image>().raycastTarget = true;
         GameManager.Instance.GetUIManager().GetInventoryUI().UpdateSlots();