+++ SurvivalGame/Assets/UIManager.cs
@@ -7,15 +7,18 @@ public class UIManager : MonoBehaviour
     [Header("Main UI GO Components Refs")]
     public GameObject inventoryUIGO;
     public GameObject quickBarUIGO;
+    public GameObject processingStationUIGO;
 
     // Script Components
     private InventoryUI inventoryUI;
     private QuickBarUI quickBarUI;
+    private ProcessingStationUI processingStationUI;
 
     void Awake()
     {
         inventoryUI = inventoryUIGO.GetComponent<InventoryUI>();
         quickBarUI = quickBarUIGO.GetComponent<QuickBarUI>();
+        processingStationUI = processingStationUIGO.GetComponent<ProcessingStationUI>();
 
         if (Instance == null) Instance = this;
         else Destroy(gameObject);
@@ -29,5 +32,10 @@ public class UIManager : MonoBehaviour
     public InventoryUI GetInventoryUI()
     {
         return inventoryUI;
+    }
+
+    public ProcessingStationUI GetProcessingStationUI()
+    {
+        return processingStationUI;
     }
 }